Fix merge conflicts for update survey
