+++ src/api/survey.js
@@ -36,3 +36,18 @@ export const surveyDelete = (id, user) => {
     }
   })
 }
+export const surveyUpdate = (id, survey, user) => {
+  return axios({
+    url: apiUrl + '/surveys/' + id,
+    method: 'PATCH',
+    headers: {
+      'Authorization': `Bearer ${user.token}`
+    },
+    data: {
+      survey: {
+        title: survey.title,
+        question: survey.question
+      }
+    }
+  })
+}
